updates to panels
