+++ app/Filament/Github/FileResource/Pages/EditFile.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Filament\Github\FileResource\Pages;
+
+use App\Filament\Github\Resources\FileResource;
+use Filament\Actions\DeleteAction;
+use Filament\Resources\Pages\EditRecord;
+
+class EditFile extends EditRecord
+{
+    protected static string $resource = FileResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            DeleteAction::make(),
+        ];
+    }
+}
