+++ app/Filament/Github/FileResource/Pages/ListFiles.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Filament\Github\FileResource\Pages;
+
+use App\Filament\Github\Resources\FileResource;
+use Filament\Actions\CreateAction;
+use Filament\Resources\Pages\ListRecords;
+
+class ListFiles extends ListRecords
+{
+    protected static string $resource = FileResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            CreateAction::make(),
+        ];
+    }
+}
