+++ app/Filament/Github/OwnerResource/Pages/EditOwner.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Filament\Github\OwnerResource\Pages;
+
+use App\Filament\Github\Resources\OwnerResource;
+use Filament\Actions\DeleteAction;
+use Filament\Resources\Pages\EditRecord;
+
+class EditOwner extends EditRecord
+{
+    protected static string $resource = OwnerResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            DeleteAction::make(),
+        ];
+    }
+}
