+++ app/Filament/Github/OwnerResource/Pages/ListOwners.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Filament\Github\OwnerResource\Pages;
+
+use App\Filament\Github\Resources\OwnerResource;
+use Filament\Actions\CreateAction;
+use Filament\Resources\Pages\ListRecords;
+
+class ListOwners extends ListRecords
+{
+    protected static string $resource = OwnerResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            CreateAction::make(),
+        ];
+    }
+}
