+++ app/Filament/Github/OwnerResource/Pages/ViewOwner.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Filament\Github\OwnerResource\Pages;
+
+use App\Filament\Github\Resources\OwnerResource;
+use Filament\Resources\Pages\ViewRecord;
+
+class ViewOwner extends ViewRecord
+{
+    protected static string $resource = OwnerResource::class;
+
+}
