+++ app/Filament/Github/Resources/RepoResource/Pages/ListRepos.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Filament\Github\Resources\RepoResource\Pages;
+
+use App\Filament\Github\Resources\RepoResource;
+use Filament\Actions\CreateAction;
+use Filament\Resources\Pages\ListRecords;
+
+class ListRepos extends ListRecords
+{
+    protected static string $resource = RepoResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            CreateAction::make(),
+        ];
+    }
+}
