--- app/Filament/Resources/CommitResource/Pages/CreateCommit.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\CommitResource\Pages;
-
-use App\Filament\Resources\CommitResource;
-use Filament\Resources\Pages\CreateRecord;
-
-class CreateCommit extends CreateRecord
-{
-    protected static string $resource = CommitResource::class;
-}
