--- app/Filament/Resources/FileResource/Pages/CreateFile.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\FileResource\Pages;
-
-use App\Filament\Resources\FileResource;
-use Filament\Resources\Pages\CreateRecord;
-
-class CreateFile extends CreateRecord
-{
-    protected static string $resource = FileResource::class;
-}
