--- app/Filament/Resources/FileResource/Pages/EditFile.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\FileResource\Pages;
-
-use App\Filament\Resources\FileResource;
-use Filament\Actions\DeleteAction;
-use Filament\Resources\Pages\EditRecord;
-
-class EditFile extends EditRecord
-{
-    protected static string $resource = FileResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            DeleteAction::make(),
-        ];
-    }
-}
