--- app/Filament/Resources/FileResource/Pages/ListFiles.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\FileResource\Pages;
-
-use App\Filament\Resources\FileResource;
-use Filament\Actions\CreateAction;
-use Filament\Resources\Pages\ListRecords;
-
-class ListFiles extends ListRecords
-{
-    protected static string $resource = FileResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            CreateAction::make(),
-        ];
-    }
-}
