--- app/Filament/Resources/OwnerResource/Pages/CreateOwner.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\OwnerResource\Pages;
-
-use App\Filament\Resources\OwnerResource;
-use Filament\Resources\Pages\CreateRecord;
-
-class CreateOwner extends CreateRecord
-{
-    protected static string $resource = OwnerResource::class;
-}
