--- app/Filament/Resources/OwnerResource/Pages/EditOwner.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\OwnerResource\Pages;
-
-use App\Filament\Resources\OwnerResource;
-use Filament\Actions\DeleteAction;
-use Filament\Resources\Pages\EditRecord;
-
-class EditOwner extends EditRecord
-{
-    protected static string $resource = OwnerResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            DeleteAction::make(),
-        ];
-    }
-}
