--- app/Filament/Resources/OwnerResource/Pages/ListOwners.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\OwnerResource\Pages;
-
-use App\Filament\Resources\OwnerResource;
-use Filament\Actions\CreateAction;
-use Filament\Resources\Pages\ListRecords;
-
-class ListOwners extends ListRecords
-{
-    protected static string $resource = OwnerResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            CreateAction::make(),
-        ];
-    }
-}
