--- app/Filament/Resources/OwnerResource/Pages/ViewOwner.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\OwnerResource\Pages;
-
-use App\Filament\Resources\OwnerResource;
-use Filament\Resources\Pages\ViewRecord;
-
-class ViewOwner extends ViewRecord
-{
-    protected static string $resource = OwnerResource::class;
-
-}
