--- app/Filament/Resources/RepoResource/Pages/CreateRepo.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\RepoResource\Pages;
-
-use App\Filament\Resources\RepoResource;
-use Filament\Resources\Pages\CreateRecord;
-
-class CreateRepo extends CreateRecord
-{
-    protected static string $resource = RepoResource::class;
-}
