--- app/Filament/Resources/RepoResource/Pages/ListRepos.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\RepoResource\Pages;
-
-use App\Filament\Resources\RepoResource;
-use Filament\Actions\CreateAction;
-use Filament\Resources\Pages\ListRecords;
-
-class ListRepos extends ListRecords
-{
-    protected static string $resource = RepoResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            CreateAction::make(),
-        ];
-    }
-}
