+++ bootstrap/providers.php
@@ -3,6 +3,7 @@
 return [
     App\Providers\AppServiceProvider::class,
     App\Providers\Filament\AdminPanelProvider::class,
+    App\Providers\Filament\GithubPanelProvider::class,
     App\Providers\GithubServiceProvider::class,
     App\Providers\HorizonServiceProvider::class,
 ];
